// src/store/modules/mutations-types.ts
export enum MutationsTypes {
  UPDATE_DISCLAIMER = "UPDATE_DISCLAIMER",
  UPDATE_PAGE = "UPDATE_PAGE",
  UPDATE_BMR = "UPDATE_BMR",
  UPDATE_LBM = "UPDATE_LBM",
  UPDATE_LBMFAT = "UPDATE_LBMFAT",
  UPDATE_TDEE = "UPDATE_TDEE",
  UPDATE_BMI = "UPDATE_BMI",
  UPDATE_MACRO = "UPDATE_MACRO",
  UPDATE_SUMMARY = "UPDATE_SUMMARY",
  UPDATE_PERFWEIGHT = "UPDATE_PERFWEIGHT",
  UPDATE_DATE = "UPDATE_DATE",
  UPDATE_BASICINFO = "UPDATE_BASICINFO",
  UPDATE_GOAL = "UPDATE_GOAL",
}